ListViewに含まれてた編集、追加、詳細画面を分割してファイルに保存。編集はEditCompanyView、追加はAddCompanyView、詳細はCompanyDetailViewとしてファイルにした。それに伴ってproject.pbxprojが変更された。
